Update to 2.1.0 - Magnetometer

+++ MetaWear.framework/Versions/A/Headers/MBLAccelerometerData.h
@@ -45,20 +45,20 @@ NS_ASSUME_NONNULL_BEGIN
 /**
  Acceleration along X axis in G's
  */
-@property (nonatomic, readonly) float x;
+@property (nonatomic, readonly) double x;
 /**
  Acceleration along Y axis in G's
  */
-@property (nonatomic, readonly) float y;
+@property (nonatomic, readonly) double y;
 /**
  Acceleration along Z axis in G's
  */
-@property (nonatomic, readonly) float z;
+@property (nonatomic, readonly) double z;
 
 /**
  RMS value of x, y, and z in G's
  */
-@property (nonatomic, readonly) float RMS;
+@property (nonatomic, readonly) double RMS;
 
 @end
